+++ src/mase/novelty/mo/NEATBreederNSGA.java
@@ -48,7 +48,7 @@ public class NEATBreederNSGA extends NEATBreeder {
 
                     @Override
                     public int compare(Individual o1, Individual o2) {
-                        return Float.compare(o2.fitness.fitness(), o1.fitness.fitness());
+                        return Double.compare(o2.fitness.fitness(), o1.fitness.fitness());
                     }
 
                 });
